Remove old tokenize and halstead code
